requirements.txt
streamlit==1.32.0
langchain==0.1.20
langchain-community==0.0.36
langchain-openai==0.1.7
langchain-pinecone==0.0.6
sentence-transformers==2.2.2
pypdf==3.17.4
pinecone-client==3.0.0
python-dotenv==1.0.0
openai==1.30.1